+++ test_autofit/graphical/hierarchical/test_embedded.py
@@ -41,21 +41,6 @@ def test_embedded_priors(
     )
 
 
-def test_hierarchical_factor_api():
-    hierarchical_factor = g.HierarchicalFactor(
-        af.GaussianPrior,
-        mean=af.GaussianPrior(
-            mean=100,
-            sigma=10
-        ),
-        sigma=af.GaussianPrior(
-            mean=10,
-            sigma=5
-        )
-    )
-    assert len(hierarchical_factor.priors) == 2
-
-
 def test_hierarchical_factor(
         centre_model
 ):
